wifi_scanner_app/requirements.txt
scapy
rich
typer
mac-vendor-lookup
python-nmap
pytest
